+++ NEWS.md
@@ -18,6 +18,8 @@
   results in a non-zero exit code (#850).
 - Improvements to waveform dumping for signals with record types or
   types with non-locally-static bounds (#851, #852).
+- The parameter for attributes with dimensions such as `'length(N)` may
+  be any integer type when `--relaxed` is passed (#862).
 - Resolved several other minor issues (#654, #854, #855, #863).
 
 ## Version 1.11.3 - 2024-02-04